checkpoint. total_var kernel randomly throwing segmentation fault. produces correct soltion when it dones't crash.

+++ test/test_toeplitz.cpp
@@ -32,15 +32,15 @@ int main(int argc, char **argv) {
     srand(101);
 
     // generate random data
-    tomocam::DArray<double> f(dim2);
-    for (int i = 0; i < f.size(); i++) f[i] = random<double>();
+    tomocam::DArray<float> f(dim2);
+    for (int i = 0; i < f.size(); i++) f[i] = random<float>();
 
     // create device array
-    tomocam::DeviceArray<double> d_f(dim2);
+    tomocam::DeviceArray<float> d_f(dim2);
     cudaMemcpy(d_f.dev_ptr(), f.begin(), f.bytes(), cudaMemcpyHostToDevice);
 
     // angles 0 - 180
-    std::vector<double> theta(nproj);
+    std::vector<float> theta(nproj);
     for (int i = 0; i < nproj; i++) theta[i] = i * M_PI / nproj;
 
     auto nugrid = tomocam::NUFFT::Grid(nproj, ncols, theta.data(), 0);
@@ -50,7 +50,7 @@ int main(int argc, char **argv) {
     auto df = psf.convolve(d_f, cudaStreamPerThread);
 
     // copy data to host
-    tomocam::DArray<double> g(df.dims());
+    tomocam::DArray<float> g(df.dims());
     cudaMemcpy(g.begin(), df.dev_ptr(), g.bytes(), cudaMemcpyDeviceToHost);
 
     // write data to hdf5
@@ -58,13 +58,13 @@ int main(int argc, char **argv) {
 
     // compute gradient using nufft
     auto dfc = tomocam::complex(d_f, cudaStreamPerThread);
-    tomocam::DeviceArray<tomocam::gpu::complex_t<double>> d_c(dim1);
-    tomocam::DeviceArray<tomocam::gpu::complex_t<double>> df2(dim2);
+    tomocam::DeviceArray<tomocam::gpu::complex_t<float>> d_c(dim1);
+    tomocam::DeviceArray<tomocam::gpu::complex_t<float>> df2(dim2);
     tomocam::NUFFT::nufft2d2(d_c, dfc, nugrid);
     tomocam::NUFFT::nufft2d1(d_c, df2, nugrid);
     auto dg2 = tomocam::real(df2, cudaStreamPerThread);
 
-    tomocam::DArray<double> g2(df2.dims());
+    tomocam::DArray<float> g2(df2.dims());
     // copy data to host
     cudaMemcpy(g2.begin(), dg2.dev_ptr(), g2.bytes(), cudaMemcpyDeviceToHost);
